Changed Camera settings and added a background
I adjusted the camera settings and added the first c#, this script is meant to control the image background. At the moment all I have done is provide a function to change the size of the sprite so it should always fill the screen you might want to change the background and do other stuff but for now this is so you can see something.
updated the scene as well